+++ classes/include/DwmGZIO.hh
@@ -216,18 +216,18 @@ namespace Dwm {
     
     //------------------------------------------------------------------------
     //!  Reads \c s from \c gzf.  Returns the number of bytes read on success,
-    //!  -1 on failure.  Since we write strings with a 32-bit unsigned length
+    //!  -1 on failure.  Since we write strings with a 64-bit unsigned length
     //!  value preceding the actual string, and always have a terminating
-    //!  NULL, this always reads at least 5 bytes on success.
+    //!  NULL, this always reads at least 9 bytes on success.
     //------------------------------------------------------------------------
     static int Read(gzFile gzf, std::string & s);
     
     //------------------------------------------------------------------------
     //!  Writes \c s to \c gzf.  Returns the number of bytes written on
-    //!  success, -1 on failure.  Note that a 32-bit value is written first,
+    //!  success, -1 on failure.  Note that a 64-bit value is written first,
     //!  holding the length of the string.  The terminating NULL is also
     //!  written.  Hence, on success this will always return a value of
-    //!  5 or greater.
+    //!  9 or greater.
     //------------------------------------------------------------------------
     static int Write(gzFile gzf, const std::string & s);
     
@@ -684,11 +684,11 @@ namespace Dwm {
         c.clear();
       int  rc = -1;
       if (gzf) {
-        uint32_t  numEntries;
+        uint64_t  numEntries;
         int  bytesRead = Read(gzf, numEntries);
         if (bytesRead == sizeof(numEntries)) {
           rc = bytesRead;
-          for (uint32_t i = 0; i < numEntries; ++i) {
+          for (uint64_t i = 0; i < numEntries; ++i) {
             typename _containerT::value_type  val;
             bytesRead = Read(gzf, val);
             if (bytesRead > 0) {
@@ -715,8 +715,8 @@ namespace Dwm {
     {
       int  rc = -1;
       if (gzf) {
-        uint32_t  numEntries = c.size();
-        uint32_t  bytesWritten = Write(gzf, numEntries);
+        uint64_t  numEntries = c.size();
+        uint64_t  bytesWritten = Write(gzf, numEntries);
         if (bytesWritten == sizeof(numEntries)) {
           rc = bytesWritten;
           if (numEntries) {
@@ -746,11 +746,11 @@ namespace Dwm {
       if (! m.empty())
         m.clear();
       if (gzf) {
-        uint32_t  numEntries;
+        uint64_t  numEntries;
         int  bytesRead = Read(gzf, numEntries);
         if (bytesRead == sizeof(numEntries)) {
           rc = bytesRead;
-          for (uint32_t i = 0; i < numEntries; ++i) {
+          for (uint64_t i = 0; i < numEntries; ++i) {
             typename _containerT::key_type  key;
             bytesRead = Read(gzf, key);
             if (bytesRead > 0) {
