+++ classes/include/DwmMacAddress.hh
@@ -156,7 +156,7 @@ namespace Dwm {
     //!  Returns the number of bytes that should be written if we call one
     //!  of the Write() members.  Should always return 6.
     //------------------------------------------------------------------------
-    uint32_t StreamedLength() const override;
+    uint64_t StreamedLength() const override;
     
     //------------------------------------------------------------------------
     //!  Prints to an ostream in 'xx:xx:xx:xx:xx:xx' format.
